+++ Frends.CSV.Parse/Frends.CSV.Parse/Definitions/Options.cs
@@ -67,4 +67,19 @@ public class Options
     /// <example>false</example>
     [DefaultValue("false")]
     public bool IgnoreQuotes { get; set; }
+
+    /// <summary>
+    /// Whether to throw an error on failure.
+    /// </summary>
+    /// <example>true</example>
+    [DefaultValue(true)]
+    public bool ThrowErrorOnFailure { get; set; } = true;
+
+    /// <summary>
+    /// Overrides the error message on failure.
+    /// </summary>
+    /// <example>CSV parsing failed: invalid input format</example>
+    [DisplayFormat(DataFormatString = "Text")]
+    [DefaultValue("")]
+    public string ErrorMessageOnFailure { get; set; } = string.Empty;
 }