+++ Frends.CSV.Parse/Frends.CSV.Parse/Definitions/Result.cs
@@ -11,31 +11,29 @@ public class Result
     /// Operation complete without errors.
     /// </summary>
     /// <example>true</example>
-    public bool Success { get; private set; }
+    public bool Success { get; internal set; }
 
     /// <summary>
     /// Processed data.
     /// </summary>
     /// <example>{ {foo, bar}, {bar, foo} }</example>
-    public List<List<object>> Data { get; private set; }
+    public List<List<object>> Data { get; internal set; }
 
     /// <summary>
     /// Headers.
     /// </summary>
     /// <example>{ foos, bars }</example>
-    public List<string> Headers { get; private set; }
+    public List<string> Headers { get; internal set; }
 
     /// <summary>
     /// The culture info used to read/write the entries. 
     /// </summary>
     /// <example>fi-FI</example>
-    public string ConfigurationCultureInfo { get; private set; }
+    public string ConfigurationCultureInfo { get; internal set; }
 
-    internal Result(bool success, List<List<object>> data, List<string> headers, string configurationCultureInfo)
-    {
-        Success = success;
-        Data = data;
-        Headers = headers;
-        ConfigurationCultureInfo = configurationCultureInfo;
-    }
+    /// <summary>
+    /// Error details. Null when Success is true.
+    /// </summary>
+    /// <example>null</example>
+    public Error Error { get; internal set; }
 }